==> 02_quality_control.ipynb <==
# Quality Control using RSeQC to verify sample quality.

Specify cram file, reference genome, area of interest, and nick name for output files.

!cramfile="HG01504.alt_bwamem_GRCh38DH.20150826.IBS.exome.cram"
!fasta="GRCh38_full_analysis_set_plus_decoy_hla.fa"
!area="chr11"
!name="HG1504_chr11"
!genemodel="hg38_RefSeq.bed" # using hg38 gene model
!readlength=101 # depends on sample

Subset data accordingly using samtools

# isolate area
!samtools view -b $cramfile $area -T $fasta > $name.bam

# sort and index resulting bam file
!samtools sort $name.bam -o sorted.$name.bam
!samtools index sorted.$name.bam

# calculate coverage of reads (and other stats)
!samtools stats sorted.$name.bam > $name.stats


Use subsetted, sorted, and indexed bam file for rseqc

# Bam file summary

Create mapping statistic summary. Look for high percentages of mapped and unique reads, and low percentages of multimapped reads.

%%bash
# cd ~/work/data
# name="0531"
# bamfile="$name.bam"
cd ~/work/data/processed
name="bioliquid_run2"
bamfile="bioliquid_run2.bam"
# bamfile="$name.sort.bam"
bam_stat.py -i $bamfile > $name_bam_stat.txt

View distribution of deletions across reads.

!pwd

%%bash
# cd ~/work/data
# name="0531"
# bamfile="0531.sort.bam"
cd ~/work/data/processed
name="bioliquid_run2"
bamfile="bioliquid_run2.bam"
readlength=101

deletion_profile.py -i $bamfile -o $name -l $readlength 
!Rscript $name.deletion_profile.r

View distribution of insertions across reads. Note that deletion distribution shows counts while insertion distribution shows percentages. 

## Insertion Deletions

!insertion_profile.py -i $bamfile -s "PE" -o $name

View GC content distribution across reads. Look for a normal distribution.

## GC Content

%%bash
# cd ~/work/data
# name="0531"
# bamfile="0531.sort.bam"
cd ~/work/data/processed
name="bioliquid_run2"
bamfile="bioliquid_run2.bam"
read_GC.py -i $bamfile -o $name

View nucleotide distribution across reads. A slight 5' bias is expected; look for even percentages across A, C, T, and G bases and low numbers of unspecified or unknown bases.

!read_NVC.py -i $bamfile -o $name -x --nx

View read quality (Phred score) distribution. Phred score = -10log10(P), where P = probability of base calling error.

## Read Quality

%%bash
cd ~/work/data
name="0531"
bamfile="0531.sort.bam"
read_quality.py -i $bamfile -o $name

View coverage across reads. A slight 5' bias is expected; look for a normal distribution.

!geneBody_coverage.py -i $bamfile -r $genemodel -o $name 

# Using Qualimap

%%bash
docker run --rm -v $PWD:/data pegi3s/qualimap qualimap bamqc -bam /data/bioliquid_run1.bam
#If you get an “Out of memory error” try
docker run --rm -v $PWD:/data pegi3s/qualimap qualimap bamqc -bam /data/bioliquid_run1.bam --java-mem-size=50G
#Where 50G is the max memory. Try using 70-80% of your total memory.

# Other Notes

%%bash
## Samtools depth
#https://www.biostars.org/p/356937/

#You can calculate the average coverage (for covered bases):
samtools depth  *bamfile*  |  awk '{sum+=$3} END { print "Average = ",sum/NR}'

#This would be average coverage for all covered regions. If you want to include regions that were not covered in this calculation you need ot use: samtools depth -a
#To get your average X coverage you would need to divide by the total size of your genome, instead of dividing by NR in the command above.
#The total size can be calculated like this:
samtools view -H *bamfile* | grep -P '^@SQ' | cut -f 3 -d ':' | awk '{sum+=$1} END {print sum}'

#Then you would redo the calculation substituting this in for NR.
#I stole all of this from: Tools To Calculate Average Coverage For A Bam File?
