==> 03_extract_reads.ipynb <==
# Extract reads

This file extracts the location of interest for each disease.

%%bash
# Input variables
run_number="run1"
chrom_dis="chr11_sca"
location=5227002
location_padding=2000000

# Setup
datadir="/mnt/aretian/genomics/nanopore/${run_number}"
run_reads="${datadir}/bioliquid_${run_number}.bam"
output="${datadir}/${run_number}_${chrom_dis}.bam"
begin=$(expr $location - $location_padding)
end=$(expr $location + $location_padding)

# Run Samtools
/home/fer/miniconda3/envs/genomics/bin/samtools view -b $run_reads "chr11:$begin-$end" > $output

# Index the file
/home/fer/miniconda3/envs/genomics/bin/samtools index $output

# Convert to sam
output="${datadir}/${run_number}_${chrom_dis}.bam"
output_sam="${datadir}/${run_number}_${chrom_dis}.sam"
/home/fer/miniconda3/envs/genomics/bin/samtools view $output > $output_sam
