==> 12_boolean_matrix.ipynb <==
# Create Boolean Matrix

This scripts creates a matrix of boolean values per read based on str presence. Example:

```
person-read   str1  str2  str3  str4  str5
p0-01         0     1     0     1     0
p0-02         1     0     0     1     0
p1-01         1     0     0     1     0
 


# Load modules

library(dplyr)
library(reshape2)

# Set parameters

run_number = 'run1'
chr='chr11'
datadir = paste0('/mnt/aretian/genomics/nanopore/',run_number)

# Load data

df <- read.csv(paste0(datadir,'/', run_number,'_',chr,'_tagged_reads.csv'), stringsAsFactors = FALSE)

# Helper functions

# instead of using length (which refers to length of STR) just use 1 to indicate it exists
agg_func <- function(l){return(1)}

unmeltdf <- function(data) {
    df <- data %>% mutate(val = 1)
    df2 <- dcast(df, read_id ~ str_id, fill = 0, value.var = 'val', fun.aggregate = agg_func)
    return(df2)
}

# Implementation

booldf <- unmeltdf(df)

# Save

write.csv(booldf, paste0(datadir,'/',run_number,'_',chr,'_bool_tagged_reads.csv'), row.names = FALSE)
